moved OrderFactory to Factory/OrderFactory
